+++ MultiLevelAuthorization/Models/RoleUser.cs
@@ -4,7 +4,7 @@ namespace MultiLevelAuthorization.Models;
 
 public class RoleUser
 {
-    public short AppId { get; set; }
+    public int AppId { get; set; }
     public Guid RoleId { get; set; }
     public Guid UserId { get; set; } 
     public Guid ModifiedByUserId { get; set; }
